requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--index-url https://mirrors.ustc.edu.cn/pypi/web/simple

absl-py==0.6.1            # via tensorflow
aiohttp==3.4.4
astor==0.7.1              # via tensorflow
async-timeout==3.0.1      # via aiohttp
attrs==18.2.0             # via aiohttp
chardet==3.0.4            # via aiohttp
gast==0.2.0               # via tensorflow
grpcio==1.18.0            # via tensorboard, tensorflow
h5py==2.8.0               # via keras, keras-applications
idna-ssl==1.1.0           # via aiohttp
idna==2.7                 # via idna-ssl, yarl
keras-applications==1.0.6  # via keras, tensorflow
keras-preprocessing==1.0.5  # via keras, tensorflow
keras-tqdm==2.0.1
keras==2.2.4
markdown==3.0.1           # via tensorboard
multidict==4.4.2          # via aiohttp, yarl
numpy==1.15.4             # via h5py, keras, keras-applications, keras-preprocessing, scipy, tensorboard, tensorflow
pillow==5.3.0
protobuf==3.6.1           # via tensorboard, tensorflow
pyyaml==3.13              # via keras
scipy==1.1.0              # via keras
six==1.11.0               # via absl-py, grpcio, h5py, keras, keras-preprocessing, protobuf, tensorboard, tensorflow
tensorboard==1.12.0       # via tensorflow
tensorflow==1.12.0
termcolor==1.1.0          # via tensorflow
tqdm==4.28.1
werkzeug==0.14.1          # via tensorboard
wheel==0.32.3             # via tensorboard, tensorflow
yarl==1.2.6               # via aiohttp
